Apply DRY principle to script files and handle changes in Ovale API.
Instead of doing a lookup of the Ovale scripts module in each
script file, do the lookup once and stash it in the addon table to
be reused. This makes it easier to adapt to changes to the Ovale
API in future releases.

Gracefully interoperate with older versions of Ovale so that this
addon will work even if someone forcibly installs an older version
of Ovale.